requirements.txt
asgiref==3.8.1
black==25.1.0
Django==4.2
django-filter==25.1
django-seed==0.3.1
django-debug-toolbar-5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.27.1
django-q==1.3
pillow==11.1.0
ipython